netlify/functions/api/requirements.txt
flask
google-generativeai
python-dotenv
gunicorn
serverless-wsgi
